+++ api_yamdb/api/urls.py
@@ -1,16 +1,13 @@
 from django.urls import include, path
 from rest_framework import routers
 
-from api import views
 from api.views import (
     CategoryViewSet,
     CommentViewSet,
     GenreViewSet,
     ReviewViewSet,
     TitleViewSet,
-    UserViewSet,
 )
-
 
 app_name = "api"
 
@@ -18,7 +15,6 @@ router_v1 = routers.DefaultRouter()
 router_v1.register("categories", CategoryViewSet, basename="categories")
 router_v1.register("genres", GenreViewSet, basename="genres")
 router_v1.register("titles", TitleViewSet, basename="titles")
-router_v1.register('users', UserViewSet, basename='users')
 router_v1.register(
     r"titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)/comments",
     CommentViewSet,
@@ -30,6 +26,4 @@ router_v1.register(
 
 urlpatterns = [
     path("v1/", include(router_v1.urls)),
-    path('v1/auth/signup/', views.signup, name="token_get"),
-    path('v1/auth/token/', views.get_token, name="token_send",)
 ]
