+++ api_yamdb/api/views.py
@@ -1,76 +1,58 @@
-import uuid
-
-from django.core.mail import send_mail
-from django.db import IntegrityError
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import filters, mixins, status, viewsets
-from rest_framework.decorators import action, api_view
-from rest_framework.filters import SearchFilter
-from rest_framework.permissions import (
-    IsAuthenticated,
-    IsAuthenticatedOrReadOnly,
-)
-from rest_framework.response import Response
-from rest_framework_simplejwt.tokens import AccessToken
-from reviews.models import Category, Genre, Review, Title, User
-
-from api_yamdb.settings import EMAIL
+from rest_framework import filters, mixins, viewsets
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from reviews.models import Category, Genre, Review, Title
 from .filters import TitleFilter
 from .permissions import (
     IsAdminModeratorAuthorPermission,
     IsAdminPermission,
-    IsOwnerOrAdmin,
 )
 from .serializers import (
     CategorySerializer,
     CommentSerializer,
     GenreSerializer,
-    MeSerializer,
     ReviewSerializer,
-    SignupSerializer,
     TitleCreateSerializer,
     TitleReadSerializer,
-    TokenSerializer,
-    UserSerializer,
 )
 
 
-class UserViewSet(viewsets.ModelViewSet):
-    """Вьюсет для объектов модели User."""
-
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-    permission_classes = (IsOwnerOrAdmin,)
-    filter_backends = (SearchFilter,)
-    search_fields = ("username",)
-    lookup_field = "username"
-    http_method_names = [
-        "get",
-        "post",
-        "patch",
-        "delete",
-        "head",
-        "options",
-        "trace",
-    ]
-
-    @action(
-        methods=["get", "patch"],
-        detail=False,
-        url_path="me",
-        permission_classes=(IsAuthenticated,),
-    )
-    def get_patch_me(self, request):
-        user = get_object_or_404(User, username=self.request.user)
-        if request.method == "GET":
-            serializer = MeSerializer(user)
-            return Response(serializer.data, status=status.HTTP_200_OK)
-        if request.method == "PATCH":
-            serializer = MeSerializer(user, data=request.data, partial=True)
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+# class UserViewSet(viewsets.ModelViewSet):
+#     """Вьюсет для объектов модели User."""
+#
+#     queryset = User.objects.all()
+#     serializer_class = UserSerializer
+#     permission_classes = (IsOwnerOrAdmin,)
+#     filter_backends = (SearchFilter,)
+#     search_fields = ("username",)
+#     lookup_field = "username"
+#     http_method_names = [
+#         "get",
+#         "post",
+#         "patch",
+#         "delete",
+#         "head",
+#         "options",
+#         "trace",
+#     ]
+#
+#     @action(
+#         methods=["get", "patch"],
+#         detail=False,
+#         url_path="me",
+#         permission_classes=(IsAuthenticated,),
+#     )
+#     def get_patch_me(self, request):
+#         user = get_object_or_404(User, username=self.request.user)
+#         if request.method == "GET":
+#             serializer = MeSerializer(user)
+#             return Response(serializer.data, status=status.HTTP_200_OK)
+#         if request.method == "PATCH":
+#             serializer = MeSerializer(user, data=request.data, partial=True)
+#             serializer.is_valid(raise_exception=True)
+#             serializer.save()
+#             return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 class ListCreateDeleteViewSet(
@@ -162,49 +144,49 @@ class ReviewViewSet(viewsets.ModelViewSet):
         )
 
 
-def validate_user_data_and_get_response(username, email):
-    serializer = UserSerializer(data={"username": username, "email": email})
-    serializer.validate({"username": username, "email": email})
-    serializer.is_valid(True)
-
-
-@api_view(["POST"])
-def signup(request):
-    serializer = SignupSerializer(data=request.data)
-    serializer.is_valid(raise_exception=True)
-    email = serializer.validated_data["email"]
-    username = serializer.validated_data["username"]
-    validate_user_data_and_get_response(username, email)
-    try:
-        user, create = User.objects.get_or_create(
-            username=username, email=email
-        )
-    except IntegrityError:
-        return Response(
-            "Такой логин или email уже существуют",
-            status=status.HTTP_400_BAD_REQUEST,
-        )
-    confirmation_code = str(uuid.uuid4())
-    user.confirmation_code = confirmation_code
-    user.save()
-    send_mail(
-        "Код подверждения",
-        confirmation_code,
-        EMAIL,
-        (email,),
-        fail_silently=False,
-    )
-    return Response(serializer.data, status=status.HTTP_200_OK)
-
-
-@api_view(["POST"])
-def get_token(request):
-    serializer = TokenSerializer(data=request.data)
-    serializer.is_valid(raise_exception=True)
-    username = serializer.validated_data["username"]
-    confirmation_code = serializer.validated_data["confirmation_code"]
-    user_base = get_object_or_404(User, username=username)
-    if confirmation_code == user_base.confirmation_code:
-        token = str(AccessToken.for_user(user_base))
-        return Response({"token": token}, status=status.HTTP_201_CREATED)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+# def validate_user_data_and_get_response(username, email):
+#     serializer = UserSerializer(data={"username": username, "email": email})
+#     serializer.validate({"username": username, "email": email})
+#     serializer.is_valid(True)
+#
+#
+# @api_view(["POST"])
+# def signup(request):
+#     serializer = SignupSerializer(data=request.data)
+#     serializer.is_valid(raise_exception=True)
+#     email = serializer.validated_data["email"]
+#     username = serializer.validated_data["username"]
+#     validate_user_data_and_get_response(username, email)
+#     try:
+#         user, create = User.objects.get_or_create(
+#             username=username, email=email
+#         )
+#     except IntegrityError:
+#         return Response(
+#             "Такой логин или email уже существуют",
+#             status=status.HTTP_400_BAD_REQUEST,
+#         )
+#     confirmation_code = str(uuid.uuid4())
+#     user.confirmation_code = confirmation_code
+#     user.save()
+#     send_mail(
+#         "Код подверждения",
+#         confirmation_code,
+#         EMAIL,
+#         (email,),
+#         fail_silently=False,
+#     )
+#     return Response(serializer.data, status=status.HTTP_200_OK)
+#
+#
+# @api_view(["POST"])
+# def get_token(request):
+#     serializer = TokenSerializer(data=request.data)
+#     serializer.is_valid(raise_exception=True)
+#     username = serializer.validated_data["username"]
+#     confirmation_code = serializer.validated_data["confirmation_code"]
+#     user_base = get_object_or_404(User, username=username)
+#     if confirmation_code == user_base.confirmation_code:
+#         token = str(AccessToken.for_user(user_base))
+#         return Response({"token": token}, status=status.HTTP_201_CREATED)
+#     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
