+++ api_yamdb/api_yamdb/settings.py
@@ -26,6 +26,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'rest_framework_simplejwt',
+    'users',
     'api',
     'reviews',
     'django_filters',
@@ -70,7 +71,7 @@ DATABASES = {
     }
 }
 
-AUTH_USER_MODEL = 'reviews.User'
+AUTH_USER_MODEL = 'users.User'
 
 AUTH_PASSWORD_VALIDATORS = [
     {
