+++ api_yamdb/reviews/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from reviews.models import Category, Comment, Genre, Review, Title, User
+from reviews.models import Category, Comment, Genre, Review, Title
 
 
 @admin.register(Genre)
@@ -41,18 +41,3 @@ class ReviewAdmin(admin.ModelAdmin):
     list_display = ("pk", "text", "score", "author", "pub_date")
     search_fields = ("text",)
     list_filter = ("pub_date",)
-
-
-@admin.register(User)
-class UserAdmin(admin.ModelAdmin):
-    list_display = (
-        "pk",
-        "username",
-        "email",
-        "first_name",
-        "last_name",
-        "bio",
-        "role",
-    )
-    search_fields = ("username", "email")
-    list_filter = ("role",)
